Detect low autocorrelation of ICA components

Low autocorrelation can be a sign of a poor quality channel or component. Often these are noisy, poor contact, or heavily contaminated with muscle noise. Low autocorrelation at a lag of 20ms is often associated with muscle noise.

ar_acf(data, ...)

# S3 method for eeg_ICA
ar_acf(data, ms = 20, plot = TRUE, verbose = TRUE,
  threshold = NULL, ...)

Arguments

data
eeg_ICA object

...
additional parameters

ms
Time lag to check ACF, in milliseconds. Defaults to 20 ms.

plot
Produce plot showing ACF and threshold for all EEG components.

verbose
Print informative messages. Defaults to TRUE.

threshold
Specify a threshold for low ACF. NULL estimates the threshold automatically.

Methods (by class)

eeg_ICA: Autocorrelation checker for eeg_ICA objects

References

Chaumon, M., Bishop, D.V., Busch, N.A. (2015). A practical guide to the selection of independent components of the electroencephalogram for artifact correction. J Neurosci Methods. Jul 30;250:47-63. doi: 10.1016/j.jneumeth.2015.02.025