browse_data {eegUtils}	R Documentation

Browse EEG data.

Description

A Shiny gadget for browsing EEG data interactively. Data can be viewed as a butterfly plot (all electrodes overlaid) or as individual traces (electrodes "stacked"). Currently, the scale cannot be manually set and is determined by the range of the viewable data.

Usage

browse_data(data, ...)

## S3 method for class 'eeg_data'
browse_data(data, sig_length = 5, n_elecs = NULL,
  downsample = FALSE, ...)

## S3 method for class 'eeg_epochs'
browse_data(data, sig_length = 5, n_elecs = NULL,
  downsample = FALSE, ...)

Arguments

data
eeg_data object to be plotted.

...
Other parameters passed to browsing functions.

sig_length
Length of signal to be plotted initially (seconds if continuous, epochs if epoched).

n_elecs
Number of electrodes to be plotted on a single screen. (not yet implemented)

downsample
Only works on eeg_data objects. Reduces size of data by only plotting every 4th point, speeding up plotting considerably.

Methods (by class)

eeg_data: Browse continuous EEG data.

eeg_epochs: Browse epoched EEG data.

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]