eeg_ar_thresh {eegUtils}	R Documentation

Simple absolute value thresholding

Description

Reject data based on a simple absolute threshold. This marks any timepoint from any electrode.

Usage

eeg_ar_thresh(data, threshold, reject = FALSE, ...)

## S3 method for class 'eeg_data'
eeg_ar_thresh(data, threshold, reject = FALSE, ...)

## S3 method for class 'eeg_epochs'
eeg_ar_thresh(data, threshold, reject = FALSE, ...)

Arguments

data
An object of class eeg_data or eeg_epochs.

threshold
In microvolts. If one value is supplied, it will be treated as a +- value.

reject
If TRUE, remove marked data immediately, otherwise mark for inspection/rejection. Defaults to FALSE.

...
Other arguments passed to eeg_ar_thresh

Methods (by class)

eeg_data: Reject data using a simple threshold.

eeg_epochs: Reject data using a simple threshold.

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]