eeg_combine {eegUtils}	R Documentation

Combine EEG objects

Description

Combine multiple eeg_epochs or eeg_data objects into a single object. Note that this does not currently perform any sort of checking for internal consistency or duplication. It simply combines the objects in the order they are passed.

Usage

eeg_combine(data, ...)

## S3 method for class 'eeg_data'
eeg_combine(data, ...)

## S3 method for class 'eeg_epochs'
eeg_combine(data, ...)

Arguments

data
An eeg_epochs object

...
additional eeg_epochs objects

Methods (by class)

eeg_data: Method for combining eeg_data objects.

eeg_epochs: Method for combining eeg_epochs objects

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]