eeg_downsample {eegUtils}	R Documentation

Downsampling EEG data

Description

Performs low-pass anti-aliasing filtering and downsamples EEG data by a specified factor. This is a wrapper for decimate from the signal package. Note that this will also adjust the event table, moving events to the nearest time remaining after downsampling

Usage

eeg_downsample(data, ...)

## S3 method for class 'eeg_data'
eeg_downsample(data, q, ...)

Arguments

data
An eeg_data object to be downsampled

...
Parameters passed to functions

q
Integer factor to downsample by

Methods (by class)

eeg_data: Downsample eeg_data objects

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]