electrode_locations {eegUtils}	R Documentation

Get standard electrode locations

Description

Joins standard electrode locations to EEG data from eegUtils internal data.

Usage

electrode_locations(data, ...)

## S3 method for class 'data.frame'
electrode_locations(data, electrode = "electrode",
  drop = FALSE, plot = FALSE, montage = NULL, ...)

## S3 method for class 'eeg_data'
electrode_locations(data, drop = FALSE, plot = FALSE,
  montage = NULL, overwrite = FALSE, ...)

Arguments

data
An EEG dataset.

...
Parameters passed to S3 methods.

electrode
The column name containing electrode names in data. (Defaults to "electrode").

drop
Should electrodes in data for which default locations are not available be dropped? (Defaults to FALSE).

plot
Plot obtained electrode locations.

montage
Name of an existing montage set. Defaults to NULL; (currently only 'biosemi64alpha' available other than default 10/20 system)

overwrite
Overwrite existing channel info. Defaults to FALSE.

Value

A tibble (or data.frame), or ggplot2 object if plot = TRUE.

Methods (by class)

data.frame: Adds standard locations to a data frame in long format

eeg_data: Adds standard locations to the chan_info field of an eeg_data object.

[Package eegUtils version 0.1.15.9001 Index]