epoch_data {eegUtils}	R Documentation

Create epochs from EEG data

Description

Creates epochs around specified event triggers. Requires data of class eeg_data. Where multiple events are specified, epochs will be created around each event.

Usage

epoch_data(data, ...)

## S3 method for class 'eeg_data'
epoch_data(data, events, time_lim = c(-1, 1), ...)

Arguments

data
Continuous data to be epoched.

...
Parameters passed to functions

events
Character vector of events to epoch around.

time_lim
Time in seconds to form epoch around the events. Defaults to one second either side.

Value

Returns an epoched object of class eeg_data

Methods (by class)

eeg_data:

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]