erp_image {eegUtils}	R Documentation

Plot ERP images

Description

Plot an ERP image from a single electrode. Smooths over a series of trials in order to make across-trial patterns more apparent.

Usage

erp_image(data, electrode = "Cz", smoothing = 10, clim = NULL)

Arguments

data
Data frame to be plotted. Requires an amplitude column.

electrode
electrode at which to generate an ERP image.

smoothing
Number of trials to smooth over when generating image

clim
Character vector of min and max values of plotting colour range. e.g. c(-5,5). Defaults to min and max.

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]