erp_scalp {eegUtils}	R Documentation

Plot ERPs on the scalp

Description

Creates an ERP figure for each electrode and layouts them on the scalp.

Usage

erp_scalp(data, electrode = "electrode", amplitude = "amplitude",
  time = "time", color = NULL, size = 0.8, baseline = NULL,
  show_guide = TRUE, montage = NULL)

Arguments

data
An EEG dataset.

electrode
Column name containing electrode names in data. Defaults to "electrode".

amplitude
Column name containing amplitudes in data. Defaults to "amplitude".

time
Column name containing time in data. Defaults to "time".

color
Variable to color lines by. If no variable is passed, only one line is drawn for each electrode.

size
Size of line(s).

baseline
Character vector of times to subtract for baseline correct.

show_guide
Should a guide be shown.

montage
Name of an existing montage set. Defaults to NULL; (currently only 'biosemi64alpha' available other than default 10/20 system)

Details

The function uses default electrode names and locations contained in the package.

Value

Returns a ggplot2 plot object.

Author(s)

Matti Vuorre, mv2521@columbia.edu

[Package eegUtils version 0.1.15.9001 Index]