interactive_scalp {eegUtils}	R Documentation

Interactive scalp maps

Description

Launches a Shiny Gadget for an interactive version of erp_scalp, allowing clicking of individual electrodes to plot them in a separate panel. In that panel, they can be averaged over or plotted as individual electrodes. Currently buggy with datasets with existing electrode co-ordinates.

Usage

interactive_scalp(data, colour = NULL, baseline = NULL, montage = NULL)

Arguments

data
An EEG dataset.

colour
Variable to colour lines by. If no variable is passed, only one line is drawn for each electrode.

baseline
Character vector of times to subtract for baseline correct.

montage
Name of an existing montage set. Defaults to NULL; (currently only 'biosemi64alpha' available other than default 10/20 system)

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]