plot_butterfly {eegUtils}	R Documentation

Create a butterfly plot from timecourse data

Description

Typically event-related potentials/fields, but could also be timecourses from frequency analyses for single frequencies. Output is a ggplot2 object. CIs not possible.

Usage

plot_butterfly(data, ...)

## Default S3 method:
plot_butterfly(data, time_lim = NULL, group = NULL,
  facet = NULL, baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, ...)

## S3 method for class 'eeg_evoked'
plot_butterfly(data, time_lim = NULL, group = NULL,
  facet = NULL, baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, ...)

## S3 method for class 'eeg_stats'
plot_butterfly(data, time_lim = NULL, group = NULL,
  facet = NULL, baseline = NULL, colourmap = NULL, legend = TRUE,
  continuous = FALSE, browse_mode = FALSE, ...)

Arguments

data
EEG dataset. Should have multiple timepoints.

...
Other parameters passed to plot_butterfly

time_lim
Character vector. Numbers in whatever time unit is used specifying beginning and end of time-range to plot. e.g. c(-.1,.3)

group
Group lines by a specificed grouping variable.

facet
Create multiple plots for a specified grouping variable.

baseline
Character vector. Times to use as a baseline. Takes the mean over the specified period and subtracts. e.g. c(-.1, 0)

colourmap
Attempt to plot using a different colourmap (from RColorBrewer). (Not yet implemented)

legend
Plot legend or not.

continuous
Is the data continuous or not (I.e. epoched)

browse_mode
Custom theme for use with browse_data.

Value

ggplot2 object showing ERPs for all electrodes overlaid on a single plot.

Methods (by class)

default: Default 'plot_butterfly' method for data.frames, eeg_data

eeg_evoked: Plot butterfly for 'eeg_evoked' objects'

eeg_stats: Butterfly plot for EEG statistics

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]