reref_eeg {eegUtils}	R Documentation

Referencing

Description

Used to reference the EEG data to specified electrode or electrodes. Defaults to average reference. Note that this stores the reference channel in the eeg_data structure; if one already exists, it is added back to the data before further referencing takes place.

Usage

reref_eeg(data, ref_chans = "average", exclude = NULL, robust = FALSE)

Arguments

data
Data to re-reference. Primarily meant for use with data of class eeg_data. Long format is expected if a data-frame is submitted.

ref_chans
Channels to reference data to. Defaults to "average" i.e. average of all electrodes in data. Character vector of channel names or numbers.

exclude
Electrodes to exclude from average reference calculation.

robust
Use median instead of mean; only applied for average reference.

Value

object of class eeg_data, re-referenced as requested.

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]