rm_baseline {eegUtils}	R Documentation

Baseline correction

Description

Used to remove the mean of a specified time period from the data. Currently only performs subtractive baseline. With a data frame, searches for "electrode" and "epoch" columns, and groups on these when found. An electrode column is always required; an epoch column is not.

Usage

rm_baseline(data, ...)

## S3 method for class 'eeg_data'
rm_baseline(data, time_lim = NULL, ...)

## S3 method for class 'eeg_epochs'
rm_baseline(data, time_lim = NULL, ...)

## S3 method for class 'data.frame'
rm_baseline(data, time_lim = NULL, ...)

Arguments

data
Data to be baseline corrected.

...
other parameters to be passed to functions

time_lim
Numeric character vector (e.g. time_lim <- c(-.1, 0)). If none given, defaults to mean of the whole of each epoch if the data is epoched, or the channel mean if the data is continuous.

Methods (by class)

eeg_data: remove baseline from continuous eeg_data

eeg_epochs: Remove baseline from eeg_epochs

data.frame: Legacy method for data.frames

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]