run_ICA {eegUtils}	R Documentation

ICA for EEG data

Description

Implements SOBI ICA. Although SOBI exists in R already, thanks to the JADE package, it doesn't respect epoch boundaries when computing correlations across different time lags. This is a port of SOBI from EEGLAB. Currently only works on epoched data.

Usage

run_ICA(data, ...)

## S3 method for class 'eeg_epochs'
run_ICA(data, method = "sobi", ...)

Arguments

data
Data frame to be ICAed.

...
Other parameters passed to function.

method
Only SOBI is currently implemented, so this is ignored.

Methods (by class)

eeg_epochs: Run ICA on an eeg_epochs object

Author(s)

Matt Craddock matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]