select_elecs {eegUtils}	R Documentation

Select electrodes from a given dataset.

Description

This is a generic function for selection of electrodes from an EEG dataset.

Usage

select_elecs(data, ...)

## Default S3 method:
select_elecs(data, electrode = NULL, keep = TRUE, ...)

## S3 method for class 'eeg_data'
select_elecs(data, electrode, keep = TRUE,
  df_out = FALSE, ...)

Arguments

data
An EEG dataset.

...
Arguments used with related methods

electrode
A character vector of electrode labels for selection or removal.

keep
Defaults to TRUE. Set to false to *remove* the selected electrodes.

df_out
Defaults to FALSE. Set to TRUE to return a dataframe rather than an eeg_data object.

Value

Data frame with only data from the chosen electrodes

eeg_data object with selected electrodes removed/kept.

Methods (by class)

default: Select electrodes from a generic data frame.

eeg_data: Select electrodes from a eeg_data object.

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]