select_times {eegUtils}	R Documentation

Select timerange

Description

Generic function for selecting specific time ranges from a given dataset. Input can be a dataframe, or an object of class eeg_data or eeg_epochs. Note this finds the closest times to those specified, so times out may not correspond exactly to requested times.

Usage

select_times(data, ...)

## Default S3 method:
select_times(data, time_lim = NULL, ...)

## S3 method for class 'eeg_data'
select_times(data, time_lim = NULL, df_out = FALSE, ...)

## S3 method for class 'eeg_epochs'
select_times(data, time_lim = NULL, df_out = FALSE,
  ...)

## S3 method for class 'eeg_evoked'
select_times(data, time_lim = NULL, df_out = FALSE,
  ...)

Arguments

data
Data from which to select

...
Further arguments passed to or from other methods.

time_lim
A character vector of two numbers indicating the time range to be selected e.g. c(min, max)

df_out
Returns a data frame rather than an object of the same type that was passed in

Value

Data frame with only data from within the specified range.

Methods (by class)

default: Default select times function

eeg_data: Select times from an eeg_data object

eeg_epochs: Select times in eeg_epoch objects

eeg_evoked: Select times in eeg_evoked objects

Author(s)

Matt Craddock, matt@mattcraddock.com

[Package eegUtils version 0.1.15.9001 Index]