Create an interpolated scalp surface

stat_scalpcontours creates an interpolated surface for an irregular set of x-y coordinates, as is typically required for a topographical EEG plot, and then calculates contours.

stat_scalpcontours(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  grid_res = 200,
  interp_limit = c("skirt", "head"),
  method = "biharmonic",
  r = NULL,
  bins = 6,
  binwidth = NULL,
  breaks = NULL,
  ...
)

Arguments

mapping

Set of aesthetic mappings created by aes(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data

The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

position

Position adjustment, either as a string naming the adjustment (e.g. "jitter" to use position_jitter), or the result of a call to a position adjustment function. Use the latter if you need to change the settings of the adjustment.

na.rm

If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend

logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes

If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

grid_res

Resolution of the interpolation grid. (Defaults to 200 points).

interp_limit

Interpolate to the "skirt" or inside the "head".

method

"biharmonic" or "gam" smooth to create interpolated surface

r

Radius of interpolated surface

bins

Number of contour bins. Overridden by binwidth.

binwidth

The width of the contour bins. Overridden by breaks.

breaks

Numeric vector to set the contour breaks. Overrides binwidth and bins. By default, this is a vector of length ten with pretty() breaks.

...

Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

See also

Author

Matt Craddock matt@mattcraddock.com