tag_events {eegUtils}	R Documentation

Tag events

Description

Give trigger events meaningful labels. Existing labels will be overwritten. Use hierarchical labelling to tag an event with multiple labels: separate labels with a "/" symbol. (e.g. "cond1" for a trigger that belongs to one condition, "cond1/cond2" for a trigger that could belong to more than one condition).

Usage

tag_events(data, ...)

## S3 method for class 'eeg_data'
tag_events(data, trigs, event_label, ...)

## S3 method for class 'eeg_epochs'
tag_events(data, trigs, event_label, ...)

Arguments

data
An object of class eeg_data or eeg_epochs

...
Parameters passed to S3 methods

trigs
Character vector of trigger numbers

event_label
Labels for the events.

Methods (by class)

eeg_data: Tag events in an eeg_data object

eeg_epochs: Tag events in an epoched dataset

Author(s)

Matt Craddock matt@mattcraddock.com

See Also

Other event handlers: list_epochs, list_events

[Package eegUtils version 0.1.15.9001 Index]