Calculate averages (e.g. ERPs) for single datasets

This function is used to create an eeg_evoked object from eeg_epochs.

eeg_average(data, ...)

# S3 method for default
eeg_average(data, ...)

# S3 method for eeg_epochs
eeg_average(data, cols = NULL, ...)

# S3 method for eeg_evoked
eeg_average(data, cols = NULL, ...)

# S3 method for eeg_tfr
eeg_average(data, cols = NULL, ...)

Arguments

data
An eeg_epochs object.

...
Other arguments passed to the averaging functions

cols
Columns from the epochs structure that the average should group on. NULL, the default, uses all columns other than the epoch column.

Methods (by class)

default: Default method for averaging EEG objects

eeg_epochs: Create evoked data from eeg_epochs

eeg_evoked: average an eeg_tfr objects over epochs.

eeg_tfr: average an eeg_tfr objects over epochs.