Plot 1d timecourse data.

Typically event-related potentials/fields, but could also be timecourses from frequency analyses for single frequencies. Averages over all submitted electrodes. Output is a ggplot2 object.

plot_timecourse(data, time_lim = NULL, group = NULL, facet = NULL,
  add_CI = FALSE, baseline = NULL, colour = NULL, electrode = NULL,
  color = NULL)

Arguments

data
EEG dataset. Should have multiple timepoints.

time_lim
Character vector. Numbers in whatever time unit is used specifying beginning and end of time-range to plot. e.g. c(-100,300)

group
(not yet implemented)

facet
Create multiple plots for a specified grouping variable.

add_CI
Add confidence intervals to the graph. Defaults to 95 percent between-subject CIs.

baseline
Character vector. Times to use as a baseline. Takes the mean over the specified period and subtracts. e.g. c(-100,0)

colour
Variable to colour lines by. If no variable is passed, only one line is drawn.

electrode
Electrode(s) to plot.

color
Alias for colour.

Value

Returns a ggplot2 plot object